Repush general

+++ app/src/Entity/Subscription.php
@@ -21,7 +21,7 @@ class Subscription
     private ?int $pdfmax = null;
 
     #[ORM\Column]
-    private ?int $price = null;
+    private ?float $price = null;
 
     #[ORM\Column(type: Types::TEXT, nullable: true)]
     private ?string $description = null;
